pyproject.toml
[project]
name = "rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "faiss-cpu>=1.12.0",
    "google-genai>=1.24.0",
    "python-dotenv>=1.1.1",
    "sentence-transformers>=5.0.0",
    "pymupdf>=1.23.0",
    "pytesseract>=0.3.10",
    "pillow>=10.0.0",
    "openai>=1.108.1"
]
